Install pip data to a non-production location
If we configure pip to install data to our real data location, then
pip helpfully tries to back it up before doing an operation. Problem
is, pip isn't actually installing data to that location, and we have
a lot of data. Misdirect pip so that it thinks an empty dir is what
it should manage.

Change-Id: I28cf2b7ce69023b7d30b9faf2a88185905ecbfee